Move the `FlateStream` from `src/core/stream.js` and into its own file
